backend/requirements.txt
fastapi
uvicorn
pydantic
python-jose
passlib[bcrypt]
python-dotenv
sqlalchemy
fastapi-mail
gunicorn